refactor: extract swallow validation
This commit does a refactoring of swallow validator for the upcoming
validations that are going to be added for the Animal model.
By introducing an animal kind validator we make sure to call only the
validator that belongs to each animal kind. So validations that assume
that the animal is a :swallow should only run the validations for the
swallow kind.

This implementation lowers the number of conditionals that we need to
make for the upcoming new validations.